fullscreen/api/document-exit-fullscreen-twice-manual.html

<!DOCTYPE html>
<title>Document#exitFullscreen() called twice</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const div = document.querySelector("div");

  document.onfullscreenchange = t.step_func(() => {
    // We are now in fullscreen.
    assert_equals(document.fullscreenElement, div);

    // Exit fullscreen twice.
    document.exitFullscreen();
    assert_equals(document.fullscreenElement, div, "fullscreenElement after first exitFullscreen()");
    const secondPromise = document.exitFullscreen();
    assert_equals(document.fullscreenElement, div, "fullscreenElement after second exitFullscreen()");

    document.onfullscreenchange = t.step_func(event => {
      assert_equals(document.fullscreenElement, null);
      assert_equals(event.target, div);

      // There will also be a second fullscreenchange event. Even though the
      // fullscreen element had already changed, the target is still the same
      // due to the "If pendingDoc's fullscreen element is not pending" logic.
      document.onfullscreenchange = t.step_func(event => {
        assert_equals(document.fullscreenElement, null);
        assert_equals(event.target, div);

        // Done, but if a promise was returned, assert that it is resolved and
        // not rejected. This test does not fail if promises aren't implemented.
        if (secondPromise) {
          secondPromise.then(t.step_func_done(), t.unreached_func("second promise rejected"));
        } else {
          t.done();
        }
      });
    });
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_request(t, div);
});
</script>
